requirements.txt
amqp==2.1.4
asgi-redis==1.1.0
asgiref==1.0.1
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
billiard==3.5.0.2
celery==4.0.2
channels==1.1.1
constantly==15.1.0
daphne==1.1.0
Django==1.10.6
django-bower==5.2.0
django-bootstrap4-datetimepicker==4.2
django-crispy-forms==1.6.1
django-model-utils==2.6.1
incremental==16.10.1
kombu==4.0.2
msgpack-python==0.4.8
olefile==0.44
Pillow==4.0.0
pytz==2016.10
redis==2.10.5
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
vine==1.1.3
zope.interface==4.3.3
